refactor: inline workspace ordering helpers, delete order.ts
The three "ordering" functions weren't real volatilities. Desktop sort
is a five-line spatial comparator that lives next to the layout read in
App.tsx; mobile order was identity (just `[...entries]`); flat order
was `entries.map(e => e.id)`. Inline them at the call sites.

Move WorkspaceSwitcherSourceEntry + buildWorkspaceEntries to model.ts
where the model lives — they shape the model's input. Delete order.ts.
