+++ packages/client/src/canvas/workspace-switcher/index.ts
@@ -7,13 +7,8 @@
 export { default } from "./WorkspaceSwitcher";
 export {
   agentBucket,
+  buildWorkspaceEntries,
   buildWorkspaceSwitcherModel,
   type WorkspaceSwitcherRepoGroup,
-} from "./model";
-export {
-  buildWorkspaceEntries,
-  desktopWorkspaceOrder,
-  flatWorkspaceOrder,
-  mobileWorkspaceOrder,
   type WorkspaceSwitcherSourceEntry,
-} from "./order";
+} from "./model";
